% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmctools.R
\docType{package}
\name{mcmctools}
\alias{mcmctools}
\alias{mcmctools-package}
\title{Useful Tools to Manipulate MCMC Objects}
\description{
Implements functions to manipulate MCMC objects (apply
    thinning, burn-in, select variables, merge chains, etc.) in the
    context of Bayesian inference.
}
