% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{mct_burnin}
\alias{mct_burnin}
\alias{mct_burnin.mcmc}
\alias{mct_burnin.mcmc.list}
\title{Burn Initials Values of the Chains}
\usage{
mct_burnin(object, burnin)

\method{mct_burnin}{mcmc}(object, burnin)

\method{mct_burnin}{mcmc.list}(object, burnin)
}
\arguments{
\item{object}{An object of class \code{\link[coda]{mcmc}} or
\code{\link[coda]{mcmc.list}}.}

\item{burnin}{The number of samples to burn.}
}
\value{
The input object \code{\link[coda]{mcmc}} or
    \code{\link[coda]{mcmc.list}} with initials samples burned.
}
\description{
Burn initials values of the chain(s) of
    \code{\link[coda]{mcmc}} or \code{\link[coda]{mcmc.list}}
    objects.
}
\examples{

data("line", package = "coda")
str(line)
str(mct_burnin(line, burnin = 5))
str(mct_burnin(line[[1L]], burnin = 100))

}
\author{
Eduardo Jr
}
