% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{mct_merge}
\alias{mct_merge}
\alias{mct_merge.mcmc.list}
\title{Merge Multiples Chains}
\usage{
mct_merge(object)

\method{mct_merge}{mcmc.list}(object)
}
\arguments{
\item{object}{An object of class \code{\link[coda]{mcmc.list}}.}
}
\value{
An object \code{\link[coda]{mcmc}}.
}
\description{
Merge the chains of a \code{\link[coda]{mcmc.list}}
    object.
}
\examples{

data("line", package = "coda")
str(line)
str(mct_merge(line))

}
\author{
Eduardo Jr
}
