% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{mct_select}
\alias{mct_select}
\alias{mct_select.default}
\alias{mct_select.mcmc}
\alias{mct_select.mcmc.list}
\title{Select Variables of a MCMC object}
\usage{
mct_select(object, variable, ...)

\method{mct_select}{default}(object, variable, ...)

\method{mct_select}{mcmc}(object, variable, ...)

\method{mct_select}{mcmc.list}(object, variable, ...)
}
\arguments{
\item{object}{An object of class \code{\link[coda]{mcmc}} or
\code{\link[coda]{mcmc.list}}.}

\item{variable}{The variables to select. Can be the column positions
(atomic integer) or column names (atomic character).}

\item{...}{Dots argument, to pass the variables separately.}
}
\value{
The input object \code{\link[coda]{mcmc}} or
    \code{\link[coda]{mcmc.list}} with the selected columns.
}
\description{
Select variables (columns) of a \code{\link[coda]{mcmc}}
    and \code{\link[coda]{mcmc.list}} objects.
}
\examples{

data("line", package = "coda")
str(line)
str(mct_select(line, "alpha"))
str(mct_select(line, "alpha", "beta"))
str(mct_select(line[[1]], "sigma"))

}
\author{
Eduardo Jr
}
