% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{mct_shinyapp}
\alias{mct_shinyapp}
\title{Interactive Interface to Analysis MCMC objects}
\usage{
mct_shinyapp(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link[coda]{mcmc}} or
\code{\link[coda]{mcmc.list}}.}

\item{...}{Others arguments passed to \code{\link[shiny]{runApp}}.}
}
\value{
NULL. This function opens the web browser with the
    interactive shiny interface.
}
\description{
Interactive interface implemented in shiny to analysis
    MCMC objects (class \code{\link[coda]{mcmc}} or
    \code{\link[coda]{mcmc.list}}).
}
\author{
Eduardo Jr
}
