% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{mct_thin}
\alias{mct_thin}
\alias{mct_thin.mcmc}
\alias{mct_thin.mcmc.list}
\title{Thin the Chains}
\usage{
mct_thin(object, thin)

\method{mct_thin}{mcmc}(object, thin)

\method{mct_thin}{mcmc.list}(object, thin)
}
\arguments{
\item{object}{An object of class \code{\link[coda]{mcmc}} or
\code{\link[coda]{mcmc.list}}.}

\item{thin}{The thining}
}
\value{
The input object \code{\link[coda]{mcmc}} or
    \code{\link[coda]{mcmc.list}} thinned.
}
\description{
Thin the chain(s) of \code{\link[coda]{mcmc}} or
    \code{\link[coda]{mcmc.list}} object.
}
\examples{

data("line", package = "coda")
str(line)
str(mct_thin(line, thin = 5))
str(mct_thin(line[[1L]], thin = 10))

}
\author{
Eduardo Jr
}
